Remove `seu` field from variant.
In the outputs from the ETL there is no column called `gnomad_seu`. This was part of the SQL load scripts but did not work with Parquet as the column was not present in the data. The field was not exposed through the API. I'm removing since that was the cause of one of the errors where the API was crashing.