Update resolution of search results to use Clickhouse.
We previously had ES loaded with a full copy of the variant index (studies and genes too, but these are small) which was used solely for search. The exact same information is in Clickhouse. I have modified the generation of ES so that we only load the fields in ES that we need for searching.

Search functions now by querying the indexes and returning a collections of IDs which match the search request. Those IDs are then resolved to objects using CH. The drawback of this approach is that we double the network time. This could be mitigated running ES in the same machine as the API as it is now so small (~5GB) that it would easily fit alongside the API which is quite slim. ES is _only_ used for a single endpoint, so it is also rarely actually hit.